move node src to parse.go
